Moved Big Sky changes from PAID_PLAN_PAID_DOMAIN_DIALOG to PAID_PLAN_IS_REQUIRED_DIALOG
Changes needed to be made to the useModalResolutionCallback hook anyway
to ensure the dialog appeared when the user had a free domain.

So switching to the PAID_PLAN_IS_REQUIRED_DIALOG dialog was better for
us anyway, since it doesn't include the green "free for a year" text
that sometimes appears.
