+++ client/my-sites/plans-features-main/components/plan-upsell-modal/paid-plan-is-required-dialog.tsx
@@ -1,8 +1,11 @@
 import { LoadingPlaceholder } from '@automattic/components';
+import { type OnboardSelect } from '@automattic/data-stores';
 import { FREE_THEME } from '@automattic/design-picker';
 import { PlanButton } from '@automattic/plans-grid-next';
+import { useDispatch, useSelect } from '@wordpress/data';
 import { useEffect, useState } from '@wordpress/element';
 import { useTranslate } from 'i18n-calypso';
+import { ONBOARD_STORE } from 'calypso/landing/stepper/stores';
 import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
 import { getHidePlanPropsBasedOnThemeType } from '../utils/utils';
 import { ButtonContainer, DialogContainer, Heading, Row } from './components';
@@ -21,6 +24,11 @@ export const PaidPlanIsRequiredDialog = ( {
 	const [ isBusy, setIsBusy ] = useState( false );
 	const isPaidTheme = selectedThemeType && selectedThemeType !== FREE_THEME;
 	const hidePlanProps = getHidePlanPropsBasedOnThemeType( selectedThemeType || '' );
+	const { setCreateWithBigSky } = useDispatch( ONBOARD_STORE );
+	const createWithBigSky = useSelect( ( select: ( key: string ) => OnboardSelect ) => {
+		const { getCreateWithBigSky } = select( ONBOARD_STORE );
+		return getCreateWithBigSky();
+	}, [] );
 
 	const getUpsellTitle = () => {
 		if ( isPaidTheme && paidDomainName ) {
@@ -31,12 +39,24 @@ export const PaidPlanIsRequiredDialog = ( {
 			return translate( 'Premium themes are only available with a paid plan' );
 		}
 
+		if ( createWithBigSky && paidDomainName ) {
+			return translate( 'Domains and our AI Website Builder are only available with a paid plan' );
+		}
+
+		if ( createWithBigSky ) {
+			return translate( 'Our AI Website Builder is only available with a paid plan' );
+		}
+
 		return translate( 'Custom domains are only available with a paid plan' );
 	};
 
 	const handleFreeDomainClick = () => {
 		setIsBusy( true );
 		onFreePlanSelected();
+
+		if ( createWithBigSky ) {
+			setCreateWithBigSky( false );
+		}
 	};
 
 	useEffect( () => {
