+++ client/my-sites/plans-features-main/components/plan-upsell-modal/paid-plan-paid-domain-dialog.tsx
@@ -1,10 +1,7 @@
 import { LoadingPlaceholder } from '@automattic/components';
-import { OnboardSelect } from '@automattic/data-stores';
 import { PlanButton } from '@automattic/plans-grid-next';
-import { useDispatch, useSelect } from '@wordpress/data';
 import { useEffect, useState } from '@wordpress/element';
 import { useTranslate } from 'i18n-calypso';
-import { ONBOARD_STORE } from 'calypso/landing/stepper/stores';
 import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
 import {
 	ButtonContainer,
@@ -24,12 +21,6 @@ export function PaidPlanPaidDomainDialog( {
 	onFreePlanSelected,
 	onPlanSelected,
 }: DomainPlanDialogProps ) {
-	const { setCreateWithBigSky } = useDispatch( ONBOARD_STORE );
-	const createWithBigSky = useSelect( ( select: ( key: string ) => OnboardSelect ) => {
-		const { getCreateWithBigSky } = select( ONBOARD_STORE );
-		return getCreateWithBigSky();
-	}, [] );
-
 	const translate = useTranslate();
 	const [ isBusy, setIsBusy ] = useState( false );
 
@@ -42,28 +33,16 @@ export function PaidPlanPaidDomainDialog( {
 	function handleFreeDomainClick() {
 		setIsBusy( true );
 		onFreePlanSelected();
-
-		if ( createWithBigSky ) {
-			setCreateWithBigSky( false );
-		}
 	}
 
-	const upsellDescription = createWithBigSky
-		? ''
-		: translate(
-				"Custom domains are only available with a paid plan. Choose annual billing and receive the domain's first year free."
-		  );
-
-	const bigSkyHeader = paidDomainName
-		? translate( 'Domains and our AI Website Builder are only available with a paid plan' )
-		: translate( 'Our AI Website Builder is only available with a paid plan' );
+	const upsellDescription = translate(
+		"Custom domains are only available with a paid plan. Choose annual billing and receive the domain's first year free."
+	);
 
 	return (
 		<DialogContainer>
 			<Heading id="plan-upsell-modal-title" shrinkMobileFont>
-				{ createWithBigSky
-					? bigSkyHeader
-					: translate( 'A paid plan is required for your domain.' ) }
+				{ translate( 'A paid plan is required for your domain.' ) }
 			</Heading>
 			<SubHeading id="plan-upsell-modal-description">{ upsellDescription }</SubHeading>
 			<ButtonContainer>
